@@ -4,15 +4,15 @@ version: 1.0.0
 homepage: https://github.com/HiroyukTamura/flutter_hls_parser
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  quiver: ^2.0.5
+  quiver: ^3.0.0
 
 dev_dependencies:
-  test: any
+  test:
   flutter_test:
     sdk: flutter
   pedantic: ^1.9.0
