@@ -1,26 +1,25 @@
-import 'package:meta/meta.dart';
 import 'segment.dart';
 import 'drm_init_data.dart';
 import 'playlist.dart';
 
 class HlsMediaPlaylist extends HlsPlaylist {
   const HlsMediaPlaylist._({
-    @required this.playlistType,
-    @required this.startOffsetUs,
-    @required this.startTimeUs,
-    @required this.hasDiscontinuitySequence,
-    @required this.discontinuitySequence,
-    @required this.mediaSequence,
-    @required this.version,
-    @required this.targetDurationUs,
-    @required this.hasEndTag,
-    @required this.hasProgramDateTime,
-    @required this.protectionSchemes,
-    @required this.segments,
-    @required this.durationUs,
-    @required String baseUri,
-    @required List<String> tags,
-    @required bool hasIndependentSegments,
+    required this.playlistType,
+    required this.startOffsetUs,
+    required this.startTimeUs,
+    required this.hasDiscontinuitySequence,
+    required this.discontinuitySequence,
+    required this.mediaSequence,
+    required this.version,
+    required this.targetDurationUs,
+    required this.hasEndTag,
+    required this.hasProgramDateTime,
+    required this.protectionSchemes,
+    required this.segments,
+    required this.durationUs,
+    required String baseUri,
+    required List<String> tags,
+    required bool hasIndependentSegments,
   }) : super(
           baseUri: baseUri,
           tags: tags,
@@ -28,21 +27,21 @@ class HlsMediaPlaylist extends HlsPlaylist {
         );
 
   factory HlsMediaPlaylist.create({
-    @required int playlistType,
-    @required int startOffsetUs,
-    @required int startTimeUs,
-    @required bool hasDiscontinuitySequence,
-    @required int discontinuitySequence,
-    @required int mediaSequence,
-    @required int version,
-    @required int targetDurationUs,
-    @required bool hasEndTag,
-    @required bool hasProgramDateTime,
-    @required DrmInitData protectionSchemes,
-    @required List<Segment> segments,
-    @required String baseUri,
-    @required List<String> tags,
-    @required bool hasIndependentSegments,
+    required int playlistType,
+    required int? startOffsetUs,
+    required int? startTimeUs,
+    required bool hasDiscontinuitySequence,
+    required int discontinuitySequence,
+    required int? mediaSequence,
+    required int? version,
+    required int? targetDurationUs,
+    required bool hasEndTag,
+    required bool hasProgramDateTime,
+    required DrmInitData? protectionSchemes,
+    required List<Segment> segments,
+    required String baseUri,
+    required List<String> tags,
+    required bool hasIndependentSegments,
   }) {
     var durationUs = segments.isNotEmpty
         ? (segments.last.relativeStartTimeUs ?? 0) +
@@ -51,7 +50,7 @@ class HlsMediaPlaylist extends HlsPlaylist {
 
     var startOffsetUsFixed = startOffsetUs ?? 0;
     if (startOffsetUsFixed < 0)
-      startOffsetUsFixed = (durationUs ?? 0) + startOffsetUs;
+      startOffsetUsFixed = (durationUs ?? 0) + startOffsetUs!;
 
     return HlsMediaPlaylist._(
       playlistType: playlistType,
@@ -85,7 +84,7 @@ class HlsMediaPlaylist extends HlsPlaylist {
 
   /// If [hasProgramDateTime] is true, contains the datetime as microseconds since epoch.
   /// Otherwise, contains the aggregated duration of removed segments up to this snapshot of the playlist.
-  final int startTimeUs;
+  final int? startTimeUs;
 
   /// Whether the playlist contains the #EXT-X-DISCONTINUITY-SEQUENCE tag.
   final bool hasDiscontinuitySequence;
@@ -94,13 +93,13 @@ class HlsMediaPlaylist extends HlsPlaylist {
   final int discontinuitySequence;
 
   /// The media sequence number of the first media segment in the playlist, as defined by #EXT-X-MEDIA-SEQUENCE, may be null if unknown.
-  final int mediaSequence;
+  final int? mediaSequence;
 
   /// The compatibility version, as defined by #EXT-X-VERSION, may be null if unknown.
-  final int version;
+  final int? version;
 
   /// The target duration in microseconds, as defined by #EXT-X-TARGETDURATION, may be null if unknown.
-  final int targetDurationUs;
+  final int? targetDurationUs;
 
   /// Whether the playlist contains the #EXT-X-ENDLIST tag.
   final bool hasEndTag;
@@ -109,11 +108,11 @@ class HlsMediaPlaylist extends HlsPlaylist {
   final bool hasProgramDateTime;
 
   /// Contains the CDM protection schemes used by segments in this playlist. Does not contain any key acquisition data. Null if none of the segments in the playlist is CDM-encrypted.
-  final DrmInitData protectionSchemes;
+  final DrmInitData? protectionSchemes;
 
   /// The list of segments in the playlist.
   final List<Segment> segments;
 
   /// The total duration of the playlist in microseconds, may be null if unknown.
-  final int durationUs;
+  final int? durationUs;
 }
