@@ -1,14 +1,12 @@
-import 'package:meta/meta.dart';
-
 abstract class HlsPlaylist {
   const HlsPlaylist({
-    @required this.baseUri,
-    @required this.tags,
-    @required this.hasIndependentSegments,
+    required this.baseUri,
+    required this.tags,
+    required this.hasIndependentSegments,
   });
 
   /// The base uri. Used to resolve relative paths.
-  final String baseUri;
+  final String? baseUri;
 
   /// The list of tags in the playlist.
   final List<String> tags;
