@@ -1,14 +1,13 @@
 import 'format.dart';
-import 'package:meta/meta.dart';
 
 class Variant {
   const Variant({
-    @required this.url,
-    @required this.format,
-    @required this.videoGroupId,
-    @required this.audioGroupId,
-    @required this.subtitleGroupId,
-    @required this.captionGroupId,
+    required this.url,
+    required this.format,
+    required this.videoGroupId,
+    required this.audioGroupId,
+    required this.subtitleGroupId,
+    required this.captionGroupId,
   });
 
   /// The variant's url.
@@ -17,19 +16,19 @@ class Variant {
   /// Format information associated with this variant.
   final Format format;
 
-  /// The video rendition group referenced by this variant, or {@code null}.
-  final String videoGroupId;
+  /// The video rendition group referenced by this variant, or [Null].
+  final String? videoGroupId;
 
-  /// The audio rendition group referenced by this variant, or {@code null}.
-  final String audioGroupId;
+  /// The audio rendition group referenced by this variant, or [Null].
+  final String? audioGroupId;
 
-  /// The subtitle rendition group referenced by this variant, or {@code null}.
-  final String subtitleGroupId;
+  /// The subtitle rendition group referenced by this variant, or [Null].
+  final String? subtitleGroupId;
 
-  /// The caption rendition group referenced by this variant, or {@code null}.
-  final String captionGroupId;
+  /// The caption rendition group referenced by this variant, or [Null].
+  final String? captionGroupId;
 
-  /// Returns a copy of this instance with the given {@link Format}.
+  /// Returns a copy of this instance with the given [Format].
   Variant copyWithFormat(Format format) => Variant(
         url: url,
         format: format,
